phishingdetection/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.11.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-Levenshtein==0.25.1
httpx==0.27.2
supabase==2.9.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
